Download a larger number of release versions
By default, GitHub's releases API returns 30 releases, which is not
always enough (for example, when trying to replicate a bug using an
older version of helm that is used on the target system).

Changes in this commit bump the number of dowloaded releases to 100,
which should suffice for now. If we would like to download all releases,
we will need to handle API pagination, which is not fun thing to do from
bash.